Fix permanently broken git cache
In some cases like VM crashes it can happen that the cache git
repository is broken but the git.Repo object doesn't detect that. In
this case the cached repo is permanently broken and the executor
cannot recover from this except taking it out of service and manually
delete all repos before starting it again.

Instead catch the 'fatal: not a git repository' message from git and
delete it completely before advancing to _ensure_cloned which will
create a fresh clone. Without the fix _ensure_cloned will still detect
a 'valid' repo and won't do anything with it.

Change-Id: Ic1f3db5afc321712ab2d42f25482e9299fd2f408