Add a sanity check for all refs returned by Gerrit
In Gerrit it is possible that a branch is created with a name that isn't
accepted by git. This leads to the situation that zuul asks for branches
that the mergers cannot check out. If that happens, config loading will
fail since the cat job will be refused by the merger due to a
non-existing branch.

Depends-On: Ic00e664152062d1fc1677e467b638ac2be99aef1
Change-Id: I6fab2bdbb1fb95412c9707f9da1b845d362600d3