Correctly show remaining time as unknown
The Python set of calculating the remaining time deals with the fact
that if not all jobs are queued, the remaining time is None. The JS
deals with this by rendering a time of 'null' as 0 minutes, which is
false.  Check for this case, and set the remaining time to 'unknown'
if the job's remaining time is null.

Change-Id: I3e1ea00084a0dc9497b9678b816883d9b984312b