Add support for authentication/STARTTLS to SMTP
Add support for user/password based authentication to the SMTP driver of
Zuul. It is now possible to use "user" and "password" as options for the
smtp driver plus "use_starttls" to encrypt the SMTP conenction.

In addition, this change also changes the way error handling is done in
the SMTP driver: A failed send attempt is now immediately reported in
the log as a warning and no error string is returned anymore since it
wasn't used on the call site.

Change-Id: I456f7a2b3e21cf72f68820064ab6acc72a369ba1