Web: don't update the status cache more than once
Once the cache expires, if multiple requestors are requesting
the status info, they all might submit gearman jobs to request
an updated status.  Instead, surround the cache check/update
with a lock so that once it expires, the first request submits
the job, all subsequent requests wait for the response and then
return the newly cached data.

Change-Id: Ifb3bfe2fe2fe5e179e53d05409897e2955a2797e