Fix node leak on job removal
If a job is removed from a project's pipeline between a node
request being submitted and fulfilled, the node request will be
accepted but not used by any job, and therefore, the nodes will
"leak" (they will be locked by Zuul and not used).  Correct this
by verifying that the item for which the request was submitted
still has the job in question when accepting the nodes; otherwise,
return them unused.

Change-Id: I4c58ea7827ea9fdc8854f29cf1d61ba0b88da029