Add coreutils to bindep on platform apk
Alpine by default comes with busybox and its version of 'du'. This
version doesn't have the '--max-depth' switch. While this can be fixed
in zuul it further unlike the gnu version of du counts the files that
are hard linked to the cache. This makes it unsuitable for zuul and we
need to require the gnu version on alpine.

Change-Id: I39032a7ea6e7b0262eed8b1c49128e64f06029d7