Optionally disable service worker in zuul-web
In some cases we need to disable the service worker which does
advanced caching. For example this is the case when using a
redirect-based authentication proxy in front of zuul. In this case it
can be that the main page is serviced by the service worker without
the option to prevent that with cache headers. When that happens the
first request that hits the server and gets redirected to a login page
is an api request. This breaks zuul-web completely until the browser
cache is cleared.

A solution to this problem is to optionally disable the service
worker. This makes it possible to disable caching of any entry pages
forcing the redirect to happen before any api requests.

This change makes it possible to disable the service worker at compile
time by setting the REACT_APP_DISABLE_SERVICE_WORKER=true variable
without the need of on-the-fly patching of zuul itself before
installation.

Change-Id: I537b2c43b556cf2c3696683bf10dd06e152ec11f