Do only one status request at a time
Currently we wait 5 seconds from the start of the request to do the
next request. On openstack the request can take more than 5 seconds so
we end up doing more requests in parallel. To fix this we should set
the timer for the next request when receiving the response data.

Change-Id: I543149f7f82cdd208cc50c30d1e29ea9a4cd453a