Retry jobs failed with MERGER_FAILURE
We sometimes get failed jobs that failed with MERGER_FAILURE. This
error is misleading to the user because it actually doesn't indicate a
merge conflict but some infrastructure related error. We already have
various retry mechanisms in place that retry most of the possible
failure causes within the executor. But catching all these places we
need to retry is difficult so we should add a safety net and
reschedule jobs that failed with MERGER_FAILURE.

Change-Id: I8844b11850c0a2cd3faddb7d8e944750c9da78ea