Exclude .keep files from .gitignore
There is a file zuul/web/static/.keep which is commited to git but
excluded from vcs by .gitignore. The problem is that some advanced IDEs
display annoying warnings related to this case.
It also looks like misconfiguration.

This patch excludes .keep files from .gitignore which means that such
files will be tracked.

Change-Id: Ib9ea9e585f723cad523d4f825c076280e2f2afe0