Fix node leak when skipping child jobs
Zuul has the possibility to skip child jobs using zuul_return. However
zuul still requests and locks nodes for these skipped jobs and never
returns them. If many jobs are skipped this can lead to a completely
wedged system. The only way to recover from this is to restart the
zuul-scheduler. This can be fixed by filtering out skipped jobs in
findJobsToRequest.

Change-Id: I69f7dd1fe142c5b4d8530407c3bca37fdd976597
Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>