Improve windows logging
Currently we don't have any stdout logs with windows at all in the
buildlog but only in the json log. This is a little bit
inconvenient. As live streaming doesn't work at the moment for windows
just don't live stream win_shell and win_command tasks but treat them
similar to localhost tasks and loop items. With this we at least get
after the fact stdout logs of these tasks.

Change-Id: Ieda6cb51dd6af0d5fb4d0a1f6f7d60d66c9fd7cb