Fix reporting ansible errors in buildlog
We had a mechanism that collects ansible errors like missing roles
that are catched early during parsing the playbook. However we had no
tests for this and this mechanism is now broken. The reason is that we
check if the first line of the syntax buffer containing the ansible
output is 'ERROR!'. However now this is the second line while the
first line tells us which ansible.cfg file it used.

To fix that just output the buffer starting with the first found line
that starts with 'ERROR!' and add a test case to it such it won't
break again.

Change-Id: Idf86bacf1117a3e95b69b45d6c70db9abb92cffc
Story: 2004289
Task: 27841