Fix "reverse" Depends-On detection with new Gerrit URL schema
The regex to detect Depends-On relations between Gerrit changes was recently
updated for the new URL schema. However because the change.uris attribute was
not also updated, getChangesDependingOn was not working, and the functionality
that this enables ("reverse" dependency processing, where a dependee change
triggers a test for an already-approved dependent) was not covered by the test.

Update the attribute and add test coverage.

Change-Id: I400445093beda390aa9ce19af31e50820f7c56df