Fix ignored but tracked .keep file
We ignore the directory zuul/web/static but want to keep the directory
itself existing. We do this by having a .keep file in there which is
excluded in the gitignore file. However we currently ignore the
complete directory zuul/web/static so git ignores the exclude
rule. Thus the .keep file is ignored but tracked which is not really
what we want. Instead ignore all files below zuul/web/static. This way
the exclude rule works and the .keep file is not ignored.

Change-Id: I3b17d1c48c6da2c5920816153a1265dbec591d38