Don't calculate priority of non-live items
Two earlier adjustments to the dynamic node priority removed
non-live items from the count of items in the queue and then
protected against the case where processOneItem attempted to
calculate the priority of a non-live item.  Instead, just avoid
calling getNodePriority from processOneItem if the item is not
live (it has no node requests), and for efficiency, remove the
safety check in getNodePriority which should no longer be
necessary.

Change-Id: Ibcf26664f3721961d8e767776f2128ec4298ffd0