Only reset working copy when needed
We perform a repo reset before each operation we do on the merger or
executor. This currently consists of fetch, reset the references,
reset and clean the working copy. The reset and clean of the working
copy in this case just picks the first found reference to reset the
repo to. While this is the wrong reference almost all the time we even
don't need a clean working copy for operations like cat, fileschanges
or repostate. Thus everything that operates on the working copy
performs a checkout first.

Especially with large repos this can take quite some time and
increases the IO load on the merger/executor. Instead we can just skip
that and ensure that the reset and clean is done properly in checkout.

Change-Id: I605b3518de2237da60f0b326c6a647a4c7ce289c