Fix items stuck in queue pending node requests
Given the following sequence:
* An item is within the active window
* A node request is submitted for a job for that item
* The active window shrinks
* A tenant reconfiguration is performed
* The node request completes

The nodes would be returned, however the record of the node request
was not removed from the item's buildset.  Once the item entered
the active window again, no further node requests would happen
for that job because of the retained node request record, and
therefore the item would be stuck in the queue.

To correct this, discard the node request record on the buildset
if we are going to return the nodes unused in that case.

Also clarify a debug message "job set" -> "job nodeset", and correct
some typos from a related unit test which were inadvertently simply
asserting true.

Change-Id: Ia6259b08ecda49f3123dae1c7d0e35dc46004561