Fix test_load_governor on large machines
The test_load_governor fakes the load to be 100 and expects the
executor to deregister. However when running the tests on a really big
machine (72 cores) the test fails because the executor permits a load
of 180 before deregistering. This can be fixed by faking the load
relative to the cpu count.

Change-Id: Ia177605356e90fc33848097f9443fc2859ac61e2