Add action to host results in zuul_json callback
Currently, this information is missing completely, although it's very
useful when somebody wants to analyze the Ansible run based on the
JSON log.

After proposing this patch also for ansible, I've learned that this
info is already visible in the original Ansible json callback:
https://github.com/ansible/ansible/pull/50853

So, I've just added this missing part to the zuul_json callback.

Change-Id: I1ee043fc1be95ec3260d3fe427653ffe8c09b8f7