Use os.path.join for git driver getGitUrl
The current code uses a format string with an unconditional injection
of a /. This can result in a double-slash which can then cause git to
be unable to clone things. That in turn can lead to an indefinite hang
of a job trying to use the repo from the git driver.

Replace the format string with os.path.join which will correctly join
the project to the baseurl whether it's a url or a bare path, and
whether or not it has a trailing slash.

Change-Id: Id0e529b916ff75d6b6df98d978d2349f6037499c