Retry jobs on winrm send_input failed
In some cases winrm tasks fail with the message 'winrm send_input
failed'. This error is unrelated to the jobs and is caused by some
underlying problems with the winrm connection or powershell on the
node. Unfortunately ansible treats this failure as a task failure
rather than like an unreachable host and thus returns with its
standard error code 2 for failed tasks so zuul doesn't catch this and
retry the job.

However this error can easily be spotted by checking the ansible
output for the string 'FATAL ERROR DURING FILE TRANSFER' which is
unique to this failure in ansible. By catching this and returning
RESULT_UNREACHABLE we can automatically retry these jobs.

Change-Id: Iceba265e9783a96b8f50dfcf269468ddee1df810