Use node v10 instead of node v8
nodejs v10 is the latest LTS release of node. Update to use it
instead of v8. One would not expect this to impact us much, but
in theory v10 is faster so we could see quicker compile times.

Update zuul-build-dashboard to run on changes to
install-js-tools.

Downloaded the 10.x version of the upstream setup script and
updated the header.

Made a meaningless change in package.json to trigger build jobs.

Change-Id: Ibfc732a9b725f8ee5f0eb51ece11692694956592