Perform per repo locking on the executor
Currently on the executor every task that operates on the git cache
repos is synchronized by a global lock so that repo updates from jobs
don't interfere with cat or merge jobs. However when dealing with
larger repos a cat or update repo job can take quite a while blocking
any other operation such as merge or job execution (via update-repo).

In order to minimize this contention we can lock per repo. This will
also help to parallelize update-repo which should improve startup
times of jobs that use many repos.

Change-Id: Iafe35c245b9dabf49d80c30b21788f849515c49f