Parallelize update threads
On systems running many jobs with many required projects the jobs can
block each other because we have a serialized single threaded queue
for repo updates. Now that we have per repo locking of the update jobs
we can parallelize them and run more than one update worker. This can
improve startup times of jobs with many required projects as well as
in situations with many jobs in startup phase.

Change-Id: I4b32d939c38111df92507eb1a87766dfad16e721