docker: add state folder
Most configuration options depend on storing state data inside
the /var/lib/zuul path which does not exist by default, causing
a lot of configurations to not work out of the box.

This patch creates the folder in the zuul base image, which will
reduce the number of options to be moved around to store state.

Change-Id: I22d05ca26324ffc02cd0b8d2a3059247a21ffdd9