web: prevent status update loop in background
When leaving the status page while it is waiting for status data,
the callback happens when the status page is not mounted,
resulting in the update loop timer not being cleared.

This change prevents that by setting the visible attribute to false
when leaving the page, so that the update is not scheduled when
the page is not visible.

Change-Id: I9f3db68bb77c52d0f477b8322bbf52bd3cc81fe1