Improve event logging in githubconnection
This moves event processing to its own class, so that it's
easier to bundle all of the data related to an event along with
an event-specific logger.

This logs the delivery ID for every line when we're preparing the
event.  It also logs the start time and queue length as well as the
end time, even on error.

Change-Id: I941a74ecbdb418cf94537ca9f8f1917a5e38dd33