Show animated progress bar in preparation phase
Show an animated progress bar in the web UI during the preparation
phase. In this phase the job already has a log URL but no 'elapsedTime'
or 'remainingTime', leading to an empty progress bar.

E.g. with large repos the prep phase can take a while during which there
is no visible feedback to the user.

Change-Id: Ia1e25e662f38f2d469eb65a16dd127dbb0414c3c