Centralize job canceling
Canceling jobs is hard because they can be in various states like
waiting for nodes, waiting for an executor or running. We further must
release the semaphore when canceling a job. We currently do this in
multiple places which can lead to hard to find bugs especially around
semaphores. Thus move the job canceling code into the scheduler at one
place and only use this for canceling jobs.

Change-Id: Iccee137909c8e99cf4e2860fb300f3100fed8dbe